+++ test/test_mpi_correctness.sh
@@ -2,18 +2,18 @@
 #SBATCH --job-name=mpi_4_nodes_correctness
 #SBATCH --output=results/mpi/mpi_4_nodes%A_%a.out
 #SBATCH --error=results/mpi/error_mpi_4_nodes%A_%a.err 
-#SBATCH --time=00:30:00 #30 min (hrs:min:sec)
+#SBATCH --time=00:10:00 #30 min (hrs:min:sec)
 #SBATCH --nodes=4
 #SBATCH --ntasks-per-node=1
 
-
 # Compilation
-rm -rf ../build
+if [ -d "../build" ]; then
+    rm -rf ../build
+fi
 mkdir ../build
 cd ../build/
 cmake -DENABLE_BENCHMARK=OFF .. 
 make -j 8 
-cd ../test/
 
 # Define path to store results
 path_results="./results/correctness/"
@@ -25,7 +25,7 @@ for i in $(seq 0 $((num_execution-1)))
 do
   arg=$((start_val * (2 ** i)))
   echo "MPI execution 1 task per node, 4 nodes, with argument: $arg"
-  mpirun -n 4 --bind-to none ../build/mpi_wf $arg >> $path_results/mpi_$arg.txt
+  mpirun -n 4 --bind-to none mpi_wf $arg >> $path_results/mpi_$arg.txt
 done
 
 # compare the element of each matrix the output file contains only one number that must to match
